+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - PropertyGrid: Fix description icon alignment for HeaderPlacement set to Above #347
 - RadioButtonList: Enhance RadioButton Control with Individual Button Enable/Disable Capability #350
 - ItemsBag: Ensure that the ItemsBag Property Descriptor does not suppress the property change notifications #354
+- ItemsBag: Propagation of IsReadOnly property to the ItemsBag Property Descriptor #369
 
 ### Changed
 - TypeHelper.IsIListIList(Type) method changed to private #343